changed it back because I'm a goober

+++ src/display.cpp
@@ -23,12 +23,12 @@ namespace SCREEN
         lcd.print(message);
     }
 
-    void print(uint8_t message)
+    void print(int message)
     {
         lcd.print(message);
     }
 
-    void setCursor(uint8_t row, uint8_t col)
+    void setCursor(int row, int col)
     {
         lcd.setCursor(row, col);
     }
